move dateRange stuff to a helper file
