+++ src/components/general/multi-date-picker/index.tsx
@@ -8,6 +8,7 @@ import {
 import { useMemo, useState } from 'react';
 import dayjs, { Dayjs } from 'dayjs';
 import updateLocale from 'dayjs/plugin/updateLocale';
+import { displayDateRanges, getDateRanges } from '@utils/helpers/dateRange';
 
 dayjs.extend(updateLocale);
 dayjs.updateLocale('en', {
@@ -78,60 +79,9 @@ function MultiDatePicker(
     );
   }
 
-  const dateRanges = useMemo(() => {
-    const dateRanges: [Dayjs, Dayjs][] = [];
-    let start: Dayjs | null = null;
-    let end: Dayjs | null = null;
-    for (const date of selection) {
-      if (start === null || end === null) {
-        start = date;
-        end = date;
-        continue;
-      }
+  const dateRanges = useMemo(() => getDateRanges(selection), [selection]);
 
-      if (date.subtract(1, 'day').isSame(end)) {
-        end = date;
-      } else {
-        dateRanges.push([start, end]);
-        start = date;
-        end = date;
-      }
-    }
-    if (start !== null && end !== null) {
-      dateRanges.push([start, end]);
-    }
-    return dateRanges;
-  }, [selection]);
-
-  const displayString = useMemo(() => {
-    let displayString = '';
-    if (dateRanges.length === 0) {
-      return 'No dates selected';
-    }
-
-    for (const [start, end] of dateRanges) {
-      if (start === end) {
-        displayString += start.format('D MMM YYYY');
-      } else {
-        const endString = end.format('D MMM YYYY');
-        const startString = start.format('D MMM YYYY');
-
-        // Show only the different part of startString
-        const startStringReversed = startString.split('').reverse();
-        const endStringReversed = endString.split('').reverse();
-        const index = startStringReversed.findIndex(
-          (c, i) => c !== endStringReversed[i]
-        );
-        const startStringPart = startStringReversed
-          .slice(index, startStringReversed.length)
-          .reverse()
-          .join('');
-        displayString += startStringPart + ' - ' + endString;
-      }
-      displayString += ', ';
-    }
-    return displayString.slice(0, -2);
-  }, [dateRanges]);
+  const displayString = useMemo(() => displayDateRanges(dateRanges), [dateRanges]);
 
   return (
     <>
